+++ R/update.R
@@ -205,10 +205,13 @@ update.brma <- function(
     fit            = fit,
     prior_list     = prior_list,
     add_parameters = .convergence_structural_parameters(object[["priors"]]),
-    max_Rhat       = object[["convergence_checks"]][["max_Rhat"]],
-    min_ESS        = object[["convergence_checks"]][["min_ESS"]],
-    max_error      = object[["convergence_checks"]][["max_error"]],
-    max_SD_error   = object[["convergence_checks"]][["max_SD_error"]]
+    max_Rhat             = object[["convergence_checks"]][["max_Rhat"]],
+    min_ESS              = object[["convergence_checks"]][["min_ESS"]],
+    max_error            = object[["convergence_checks"]][["max_error"]],
+    max_SD_error         = object[["convergence_checks"]][["max_SD_error"]],
+    check_indicators     = isTRUE(object[["convergence_checks"]][["check_indicators"]]),
+    monitor              = object[["convergence_checks"]][["monitor"]],
+    allow_not_assessable = isTRUE(object[["convergence_checks"]][["allow_not_assessable"]])
   )
 
   fit[["converged"]]     <- check_fit